third_party/WebKit/LayoutTests/external/wpt/innerText/setter-expected.txt

This is a testharness.js-based test.
Found 126 tests; 80 PASS, 46 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Simplest possible test
PASS Simplest possible test, detached
PASS Newlines convert to <br> in non-white-space:pre elements
PASS Newlines convert to <br> in non-white-space:pre elements, detached
FAIL Newlines convert to <br> in <pre> elementassert_equals: expected "abc<br>def" but got "abc\ndef"
PASS Newlines convert to <br> in <pre> element, detached
FAIL Newlines convert to <br> in <textarea> elementassert_equals: expected "abc<br>def" but got "abc\ndef"
PASS Newlines convert to <br> in <textarea> element, detached
FAIL Newlines convert to <br> in white-space:pre elementassert_equals: expected "abc<br>def" but got "abc\ndef"
PASS Newlines convert to <br> in white-space:pre element, detached
PASS CRs convert to <br> in non-white-space:pre elements
PASS CRs convert to <br> in non-white-space:pre elements, detached
FAIL CRs convert to <br> in <pre> elementassert_equals: expected "abc<br>def" but got "abc\ndef"
PASS CRs convert to <br> in <pre> element, detached
PASS Newline/CR pair converts to <br> in non-white-space:pre element
PASS Newline/CR pair converts to <br> in non-white-space:pre element, detached
FAIL Newline/newline pair converts to two <br>s in non-white-space:pre elementassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Newline/newline pair converts to two <br>s in non-white-space:pre element, detachedassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL CR/CR pair converts to two <br>s in non-white-space:pre elementassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL CR/CR pair converts to two <br>s in non-white-space:pre element, detachedassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL CRs convert to <br> in white-space:pre elementassert_equals: expected "abc<br>def" but got "abc\ndef"
PASS CRs convert to <br> in white-space:pre element, detached
PASS < preserved
PASS < preserved, detached
PASS > preserved
PASS > preserved, detached
PASS & preserved
PASS & preserved, detached
PASS " preserved
PASS " preserved, detached
PASS ' preserved
PASS ' preserved, detached
PASS innerText not supported on SVG elements
PASS innerText not supported on SVG elements, detached
PASS innerText not supported on MathML elements
PASS innerText not supported on MathML elements, detached
PASS Null characters preserved
PASS Null characters preserved, detached
PASS Tabs preserved
PASS Tabs preserved, detached
PASS Leading whitespace preserved
PASS Leading whitespace preserved, detached
PASS Trailing whitespace preserved
PASS Trailing whitespace preserved, detached
PASS Whitespace not compressed
PASS Whitespace not compressed, detached
FAIL Existing text deletedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
FAIL Existing text deleted, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS Existing <br> deleted
PASS Existing <br> deleted, detached
PASS Assigning the empty string
PASS Assigning the empty string, detached
PASS Assigning null
PASS Assigning null, detached
PASS Assigning undefined
PASS Assigning undefined, detached
FAIL Start with CRassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Start with CR, detachedassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Start with LFassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Start with LF, detachedassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Start with CRLFassert_not_equals: Should not have empty text nodes got disallowed value ""
FAIL Start with CRLF, detachedassert_not_equals: Should not have empty text nodes got disallowed value ""
PASS End with CR
PASS End with CR, detached
PASS End with LF
PASS End with LF, detached
PASS End with CRLF
PASS End with CRLF, detached
PASS innerText on <area> element
FAIL innerText on <area> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <base> element
FAIL innerText on <base> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <basefont> element
FAIL innerText on <basefont> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <bgsound> element
FAIL innerText on <bgsound> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <br> element
FAIL innerText on <br> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <col> element
FAIL innerText on <col> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <embed> element
FAIL innerText on <embed> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <frame> element
FAIL innerText on <frame> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <hr> element
FAIL innerText on <hr> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <image> element
FAIL innerText on <image> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <img> element
FAIL innerText on <img> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <input> element
FAIL innerText on <input> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <keygen> element
FAIL innerText on <keygen> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <link> element
FAIL innerText on <link> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <menuitem> element
FAIL innerText on <menuitem> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <meta> element
FAIL innerText on <meta> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <param> element
FAIL innerText on <param> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <source> element
FAIL innerText on <source> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <track> element
FAIL innerText on <track> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <wbr> element
FAIL innerText on <wbr> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <colgroup> element
FAIL innerText on <colgroup> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <frameset> element
FAIL innerText on <frameset> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <head> element
FAIL innerText on <head> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <html> element
FAIL innerText on <html> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <table> element
FAIL innerText on <table> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <tbody> element
FAIL innerText on <tbody> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <tfoot> element
FAIL innerText on <tfoot> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <thead> element
FAIL innerText on <thead> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
PASS innerText on <tr> element
FAIL innerText on <tr> element, detachedassert_not_equals: Child should be a *new* text node got disallowed value Text node "abc"
Harness: the test ran to completion.
